Add files via upload
The Jupyter notebook is designed to analyze the data collected by various fitness trackers, such as Apple Watch and Fitbit. It displays various visualizations and comparisons of the data. Through its code and visualizations, users can learn more about the reliability and accuracy of the data that these devices collect.